Added XmlTag support and subclassable stringmgr support. Allowed casting to subtypes of SWModule and SWKey without going through intermediate stages. (e.g. SWModule -> RawLD4, instead of SWModule -> SWLD -> RawLD)
Wrapped setLocale and builtin_abbrevs, as well as added methods on 
the abbreviation to allow iterating through them. Set sbook and abbrevs
members to be immutable to remove warnings.

Wrapped some StringList's directly, in addition to the Vector methods
For example, SWMgr.getGlobalOptions can be called directly, rather than using
SWMgr.getGlobalOptionsVector.

Fixed installmgr wrapping so that it works again.

Fixed python setup to be able to be run from the python directory properly
Updated version number to 1.5.11



git-svn-id: https://svn.crosswire.org/svn/sword/trunk@2158 bcd7d363-81e1-0310-97ec-a550e20fc99c