Cleaned up some CMake problems.
I took out a bunch of extraneous CMake output messages and bound up
several messages into more standardized output formats.

I also eliminated a configure problem where the bindings configure script
would be called even if the user had not selected bindings to be built.
This could cause configure to fail if SWIG was not found when the user did
not wish bindings to be built.


git-svn-id: https://svn.crosswire.org/svn/sword/trunk@2592 bcd7d363-81e1-0310-97ec-a550e20fc99c